Move backtracking to utils folder
